+++ src/diffusions/gbm.rs
@@ -4,11 +4,11 @@ use crate::{
 };
 use ndarray::Array1;
 
-pub fn gbm(mu: f64, sigma: f64, n: usize, x0: Option<f64>, t: Option<f64>) -> Vec<f64> {
+pub fn gbm(mu: f32, sigma: f32, n: usize, x0: Option<f32>, t: Option<f32>) -> Vec<f32> {
   let gn = gn(n - 1, Some(t.unwrap_or(1.0)));
-  let dt = t.unwrap_or(1.0) / n as f64;
+  let dt = t.unwrap_or(1.0) / n as f32;
 
-  let mut gbm = Array1::<f64>::zeros(n);
+  let mut gbm = Array1::<f32>::zeros(n);
   gbm[0] = x0.unwrap_or(100.0);
 
   for i in 1..n {
@@ -19,21 +19,21 @@ pub fn gbm(mu: f64, sigma: f64, n: usize, x0: Option<f64>, t: Option<f64>) -> Ve
 }
 
 pub fn fgbm(
-  hurst: f64,
-  mu: f64,
-  sigma: f64,
+  hurst: f32,
+  mu: f32,
+  sigma: f32,
   n: usize,
-  x0: Option<f64>,
-  t: Option<f64>,
-) -> Vec<f64> {
+  x0: Option<f32>,
+  t: Option<f32>,
+) -> Vec<f32> {
   if !(0.0..1.0).contains(&hurst) {
     panic!("Hurst parameter must be in (0, 1)")
   }
 
   let fgn = FgnFft::new(hurst, n - 1, t, None).sample();
-  let dt = t.unwrap_or(1.0) / n as f64;
+  let dt = t.unwrap_or(1.0) / n as f32;
 
-  let mut fgbm = Array1::<f64>::zeros(n);
+  let mut fgbm = Array1::<f32>::zeros(n);
   fgbm[0] = x0.unwrap_or(100.0);
 
   for i in 1..n {
