+++ src/statistics/fractal_dim.rs
@@ -1,6 +1,6 @@
 use linreg::linear_regression;
 
-pub fn higuchi_fd(x: &[f64], kmax: usize) -> f64 {
+pub fn higuchi_fd(x: &[f32], kmax: usize) -> f32 {
   let n_times = x.len();
 
   let mut lk = vec![0.0; kmax];
@@ -12,19 +12,19 @@ pub fn higuchi_fd(x: &[f64], kmax: usize) -> f64 {
 
     for m in 0..k {
       let mut ll = 0.0;
-      let n_max = ((n_times - m - 1) as f64 / k as f64).floor() as usize;
+      let n_max = ((n_times - m - 1) as f32 / k as f32).floor() as usize;
 
       for j in 1..n_max {
         ll += (x[m + j * k] - x[m + (j - 1) * k]).abs();
       }
 
-      ll /= k as f64;
-      ll *= (n_times - 1) as f64 / (k * n_max) as f64;
+      ll /= k as f32;
+      ll *= (n_times - 1) as f32 / (k * n_max) as f32;
       lm[m] = ll;
     }
 
-    lk[k - 1] = lm.iter().sum::<f64>() / k as f64;
-    x_reg[k - 1] = (1.0 / k as f64).ln();
+    lk[k - 1] = lm.iter().sum::<f32>() / k as f32;
+    x_reg[k - 1] = (1.0 / k as f32).ln();
     y_reg[k - 1] = lk[k - 1].ln();
   }
